+++ CHANGELOG.md
@@ -87,6 +87,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - This allows these endpoints to be accessed through port `ETV_STREAMING_PORT` (default `8409`)
   - This only matters if you configured `ETV_UI_PORT` to be a different value, which makes UI endpoints inaccessible on the streaming port
 - Update Plex movie/other video plot ("summary") during library deep scan
+- Fix compatibility with ffmpeg 7.2+ when using NVIDIA accel and 10-bit source content
 
 ## [25.2.0] - 2025-06-24
 ### Added